requirements.txt
flask==3.0.0
python-dotenv==1.0.0
google-generativeai==0.5.0
firebase-admin==6.4.0
flask-limiter==3.5.0
flask-cors==4.0.0
flask-compress==1.14
flask-caching==2.1.0
flask-wtf==1.2.1
requests==2.31.0
bleach==6.1.0
pytest==8.0.0